Account

+++ src/main/java/com/project/api/entity/Challenge.java
@@ -38,9 +38,9 @@ public class Challenge {
 
     private String description;
 
-    private double targetAmount;
+    private Long targetAmount;
 
-    private double currentProgress;
+    private Long currentProgress;
 
     private LocalDateTime startDate;
